requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.1.1
django-admin-extra-buttons==1.5.8
django-import-export==4.1.1
frozendict==2.4.6
gevent==24.2.1
greenlet==3.1.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
logzero==1.7.0
lxml==5.3.0
multitasking==0.0.11
numpy==2.1.1
packaging==24.1
pandas==2.2.3
peewee==3.17.7
platformdirs==4.3.6
psycopg2-binary==2.9.9
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
smartapi-python==1.5.1
soupsieve==2.6
sqlparse==0.5.1
ta==0.11.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.7.0
yfinance==0.2.46
zope.event==5.0
zope.interface==7.0.3